schools

+++ app/controllers/classRooms.js
@@ -4,14 +4,14 @@ var logger = require('winston'),
 
 // Routes
 module.exports = function(app) {
-    app.post('/addClassRoom', add);
-    app.post('/updateClassRoom', update);
-    app.post('/removeClassRoom',remove);
-    app.get('/getClassRooms', getAll); 
+    app.post('/addClassRoom', addClassRoom);
+    app.post('/updateClassRoom', updateClassRoom);
+    app.post('/removeClassRoom',removeClassRoom);
+    app.get('/getClassRooms', getClassRooms); 
 };
 
 // Controlls
-add = function(req, res) {
+var addClassRoom = function(req, res) {
     if (req.session.user) {
         var newRoom = req.body;
     	var query = "INSERT into classRooms (name, capacity, description) values (?, ?, ?)";
@@ -22,7 +22,10 @@ add = function(req, res) {
             		logger.error("Такая аудитория уже существует:", newRoom.name)
                 	res.send({error: "Такая аудитория уже существует"});
                 }
-                else logger.error(err);
+                else {
+                    logger.error(query, ":", err);
+                    res.send({error: "Ошибка сервера. Выполнить операцию не удалось."});
+                }
             }
             else {
             	logger.info("Добавлена аудитория:", newRoom.name);
@@ -32,7 +35,7 @@ add = function(req, res) {
     }
 }
 
-update = function(req, res) {
+var updateClassRoom = function(req, res) {
     if (req.session.user) {
         var room = req.body;
         var query = "UPDATE classRooms set (name, capacity, description) = (?, ?, ?) WHERE id = " + room.id;
@@ -42,7 +45,10 @@ update = function(req, res) {
                     logger.error("Такая аудитория уже существует:", room.name)
                     res.send({error: "Такая аудитория уже существует"});
                 }
-                else logger.error(err);
+                else {
+                    logger.error(query, ":", err);
+                    res.send({error: "Ошибка сервера. Выполнить операцию не удалось."});
+                }
             }
             else {
                 logger.info("Обновлена аудитория:", room.name);
@@ -52,34 +58,40 @@ update = function(req, res) {
     }
 }
 
-remove = function(req, res) {
-    var room = req.body;
-	var query = "DELETE FROM classRooms where id = " + room.id ;
-	
-    database.run(query, function(err, row) {
-        if (err) {
-            logger.error(err);
-            res.send(err);
-        }
-        else {
-        	logger.info("Удалена аудитория:", room.name);
-        }
-        res.send(true);
-    })
+var removeClassRoom = function(req, res) {
+    if (req.session.user) {
+        var room = req.body;
+    	var query = "DELETE FROM classRooms where id = " + room.id ;
+    	
+        database.run(query, function(err, row) {
+            if (err) {
+                logger.error(query, ":", err);
+                res.send({error: "Ошибка сервера. Выполнить операцию не удалось."});
+            }
+            else {
+            	logger.info("Удалена аудитория:", room.name);
+                res.send(true);
+            }
+        })
+    }
 }
 
-getAll = function(req, res) {	
-	var rooms = [];
-	
-	database.all("SELECT id, name, capacity, description FROM classRooms", function(err, rows) {
-	    if (err) {
-            logger.error(err);
-	    	res.send(err);
-	    }
-	    rows.forEach(function (row) {
-	    	room = new classRoom(row);
-	    	rooms.push(room);
-	    });
-	    res.send(rooms);
-	});
+var getClassRooms = function(req, res) {
+    if (req.session.user) {
+    	var rooms = [];
+    	
+    	database.all("SELECT id, name, capacity, description FROM classRooms", function(err, rows) {
+    	    if (err) {
+                logger.error("GET all FROM classRooms", err);
+    	    	res.send({error: "Ошибка сервера. Выполнить операцию не удалось."});
+    	    }
+            else {
+                rows.forEach(function (row) {
+                    room = new classRoom(row);
+                    rooms.push(room);
+                });
+                res.send(rooms);
+            }
+    	});
+    }
 }
